Start serving the plugin API prior to registering with kubelet
This is the recommended order when starting a plugin.
Additionally, not starting the server prior to registration
causes issues starting with K8s 1.25. It is unclear exactly
what changed in 1.25 that brings this issue to surface, but
it is possible kubelet now implicitly assumes a plugin
is starting to serve the API upon receiving a registration
request.

Without this change, we see the following kubelet logs:

"Got registration request from device plugin with resource" resourceName="nvidia.com/***"
"Unable to connect to device plugin client with socket path" err="failed to dial device plugin:
context deadline exceeded" path="/var/lib/kubelet/device-plugins/kubevirt-***.sock"

Signed-off-by: Christopher Desiniotis <cdesiniotis@nvidia.com>